Rename(#42): Member, SocialMember 위치 변경
