+++ application/wypl-core/src/main/java/com/wypl/wyplcore/schedule/service/ScheduleService.java
@@ -1,5 +1,9 @@
 package com.wypl.wyplcore.schedule.service;
 
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.wypl.googleoauthclient.domain.AuthMember;
 import com.wypl.jpacalendardomain.calendar.domain.Calendar;
 import com.wypl.jpacalendardomain.calendar.domain.Schedule;
 import com.wypl.jpacalendardomain.calendar.domain.ScheduleInfo;
@@ -7,13 +11,11 @@ import com.wypl.jpacalendardomain.calendar.mapper.ScheduleInfoMapper;
 import com.wypl.jpacalendardomain.calendar.mapper.ScheduleMapper;
 import com.wypl.jpacalendardomain.calendar.repository.ScheduleInfoRepository;
 import com.wypl.jpacalendardomain.calendar.repository.ScheduleRepository;
-import com.wypl.jpamemberdomain.member.Member;
-import com.wypl.wyplcore.auth.domain.AuthMember;
+import com.wypl.jpamemberdomain.member.domain.Member;
 import com.wypl.wyplcore.schedule.data.request.ScheduleCreateRequest;
 import com.wypl.wyplcore.schedule.data.response.ScheduleInfoCreateResponse;
+
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 
 @Service
@@ -29,7 +31,7 @@ public class ScheduleService {
         Calendar foundCalendar = null;  // FIXME: scheduleInfoRequest의 calendarId로 찾는다. foundCalendar 엔티티 검증 필요.
         Member foundMember = null; // FIXME: member 엔티티 검증 필요.
 
-        ScheduleInfo scheduleInfo = ScheduleInfoMapper.toJpaScheduleInfo(foundCalendar, authMember);
+        ScheduleInfo scheduleInfo = ScheduleInfoMapper.toJpaScheduleInfo(foundCalendar, authMember.id());
         Schedule schedule = ScheduleMapper.toJpaSchedule(scheduleCreateRequest, scheduleInfo);
 
         ScheduleInfo savedScheduleInfo = scheduleInfoRepository.save(scheduleInfo);
