+++ domain/jpa-calendar-domain/src/main/java/com/wypl/jpacalendardomain/calendar/domain/MemberCalendar.java
@@ -5,7 +5,7 @@ import org.hibernate.annotations.SQLRestriction;
 
 import com.wypl.common.Color;
 import com.wypl.jpacalendardomain.calendar.data.InviteStatus;
-import com.wypl.jpamemberdomain.member.Member;
+import com.wypl.jpamemberdomain.member.domain.Member;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
