+++ domain/jpa-member-domain/src/main/java/com/wypl/jpamemberdomain/member/domain/SocialMember.java
@@ -1,4 +1,6 @@
-package com.wypl.jpamemberdomain.member;
+package com.wypl.jpamemberdomain.member.domain;
+
+import com.wypl.jpamemberdomain.member.OauthProvider;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
